+++ src/app/Http/Requests/User/Favorite/IndexFavoriteRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Requests\User;
+namespace App\Http\Requests\User\Favorite;
 
 use App\Http\Requests\Base\ApiRequest;
